refactor: Simplify date conversion logic in CalendarConverter
- Delegate Gregorian to Jalali and Jalali to Gregorian conversions to specialized converter classes for improved maintainability.
- Update references to JalaliDate model to use the new import path.
- Enhance the documentation to clarify the purpose of the CalendarConverter class.
